+++ tests/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_executable(bfs_tests
-    test_boost_mpl_graph_bfs_route_finder.cpp
-    test_constexpr_graph_bfs_route_finder.cpp
+    test_boost_mpl_graph_shortest_path.cpp
+    test_constexpr_graph_shortest_path.cpp
 )
 
 target_link_libraries(bfs_tests
